more work on builtins, restructuring

+++ expandvalue.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   expandvalue.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: adinari <adinari@student.42.fr>            +#+  +:+       +#+        */
+/*   By: slakner <slakner@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/17 20:18:59 by adinari           #+#    #+#             */
-/*   Updated: 2022/11/18 06:06:11 by adinari          ###   ########.fr       */
+/*   Updated: 2022/11/18 19:10:24 by slakner          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ char	*value_expand(char **envp, char *var)
 {
 	int		j;
 	char	*value;
-	int		len;
+	size_t	len;
 	char	*tmp;
 
 	len = 0;
